[infrastructure] improvement to cake
