+++ cake/BuildContext.cs
@@ -8,20 +8,45 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.IO;
+using System.Text;
 using Cake.Common.Tools.DotNet;
 using Cake.Common.Tools.DotNet.Build;
 using Cake.Common.Tools.DotNet.MSBuild;
 using Cake.Common.Tools.DotNet.Run;
 using Cake.Common.Tools.DotNet.Test;
 using Cake.Core;
-using Cake.Core.IO;
 using Cake.Frosting;
 using Polly;
 using Path = System.IO.Path;
 
 public class BuildContext : FrostingContext
 {
+
+    public void UpdateApaxVersion(string file, string version)
+    {
+        var sb = new StringBuilder();
+        foreach (var line in System.IO.File.ReadLines(file))
+        {
+            var newLine = line;
+
+            if (line.Trim().StartsWith("version"))
+            {
+                var semicPosition = line.IndexOf(":");
+                var lenght = line.Length - semicPosition;
+
+                newLine = $"{line.Substring(0, semicPosition)} : '{version}'";
+            }
+            sb.AppendLine(newLine);
+        }
+
+        System.IO.File.WriteAllText(file, sb.ToString());
+    }
     public string Artifacts  => Path.Combine(Environment.WorkingDirectory.FullPath, "..//artifacts//");
+
+    public string ArtifactsApax => EnsureFolder(Path.Combine(Artifacts, "apax"));
+    public string ArtifactsNugets => EnsureFolder(Path.Combine(Artifacts, "nugets"));
+
+    public string PackableNugetsSlnf => Path.Combine(RootDir, "ix.framework-packable-only.slnf");
 
     public string WorkDirName => Environment.WorkingDirectory.GetDirectoryName();
 
@@ -92,81 +117,14 @@ public class BuildContext : FrostingContext
     {
         return Path.Combine(Path.Combine(RootDir, library.folder), "ctrl", "apax.yml");
     }
-}
 
-public static class Apax
-{
-    public static void ApaxInstall(this BuildContext context, (string folder, string name) lib)
+    public string EnsureFolder(string path)
     {
-        context.ProcessRunner.Start(Helpers.GetApaxCommand(), new ProcessSettings()
+        if (!Directory.Exists(path))
         {
-            Arguments = "install -L",
-            WorkingDirectory = context.GetAxFolder(lib),
-            RedirectStandardOutput = false,
-            RedirectStandardError = false,
-            Silent = false
-        }).WaitForExit();
-    }
+            Directory.CreateDirectory(path);
+        }
 
-    public static void ApaxClean(this BuildContext context, (string folder, string name) lib)
-    {
-        context.ProcessRunner.Start(Helpers.GetApaxCommand(), new ProcessSettings()
-        {
-            Arguments = "clean",
-            WorkingDirectory = context.GetAxFolder(lib),
-            RedirectStandardOutput = false,
-            RedirectStandardError = false,
-            Silent = false
-        }).WaitForExit();
-    }
-
-    public static void ApaxBuild(this BuildContext context, (string folder, string name) lib)
-    {
-        context.ProcessRunner.Start(Helpers.GetApaxCommand(), new ProcessSettings()
-        {
-            Arguments = "build",
-            WorkingDirectory = context.GetAxFolder(lib),
-            RedirectStandardOutput = false,
-            RedirectStandardError = false,
-            Silent = false
-        }).WaitForExit();
-    }
-
-    public static void ApaxPack(this BuildContext context, (string folder, string name) lib)
-    {
-        context.ProcessRunner.Start(Helpers.GetApaxCommand(), new ProcessSettings()
-        {
-            Arguments = "pack",
-            WorkingDirectory = context.GetAxFolder(lib),
-            RedirectStandardOutput = false,
-            RedirectStandardError = false,
-            Silent = false
-        }).WaitForExit();
-    }
-
-    public static void ApaxTest(this BuildContext context, (string folder, string name) lib)
-    {
-        context.ProcessRunner.Start(Helpers.GetApaxCommand(), new ProcessSettings()
-        {
-            Arguments = "test",
-            WorkingDirectory = context.GetAxFolder(lib),
-            RedirectStandardOutput = false,
-            RedirectStandardError = false,
-            Silent = false
-        }).WaitForExit();
-    }
-
-    public static void ApaxPublish(this BuildContext context, (string folder, string name) lib)
-    {
-        var libraryFolder = Path.Combine(Path.Combine(context.RootDir, lib.folder), "ctrl");
-        var packageFile = $"ix-ax-{lib.name}-{GitVersionInformation.SemVer}.apax.tgz";
-        context.ProcessRunner.Start(Helpers.GetApaxCommand(), new ProcessSettings()
-        {
-            Arguments = $"publish -p {packageFile} -r  https://npm.pkg.github.com",
-            WorkingDirectory = libraryFolder,
-            RedirectStandardOutput = false,
-            RedirectStandardError = false,
-            Silent = false
-        }).WaitForExit();
+        return path;
     }
 }
