+++ cake/Program.cs
@@ -131,25 +131,7 @@ public sealed class TestsTask : FrostingTask<BuildContext>
 [IsDependentOn(typeof(TestsTask))]
 public sealed class CreateArtifactsTask : FrostingTask<BuildContext>
 {
-    private void UpdateApaxVersion(string file, string version)
-    {
-        var sb = new StringBuilder();
-        foreach (var line in System.IO.File.ReadLines(file))
-        {
-            var newLine = line;
-
-            if (line.Trim().StartsWith("version"))
-            {
-                var semicPosition = line.IndexOf(":");
-                var lenght = line.Length - semicPosition;
-
-                newLine = $"{line.Substring(0, semicPosition)} : '{version}'";
-            }
-            sb.AppendLine(newLine);
-        }
-
-        System.IO.File.WriteAllText(file, sb.ToString());
-    }
+   
 
     public override void Run(BuildContext context)
     {
@@ -159,33 +141,27 @@ public sealed class CreateArtifactsTask : FrostingTask<BuildContext>
             return;
         }
 
+        PackApax(context);
+        PackNugets(context);
+    }
+
+    private static void PackApax(BuildContext context)
+    {
         context.Libraries.ToList().ForEach(lib =>
         {
-            UpdateApaxVersion(context.GetApaxFile(lib), GitVersionInformation.SemVer);
+            context.UpdateApaxVersion(context.GetApaxFile(lib), GitVersionInformation.SemVer);
             context.ApaxPack(lib);
+            context.ApaxCopyArtifacts(lib);
         });
-
-        PackPackages(context, Path.Combine(context.RootDir, "ix.framework-packable-only.slnf"));
     }
 
-    private static void PackTemplatePackages(BuildContext context, string solutionToPack)
-    {
-        context.DotNetPack(solutionToPack,
-            new Cake.Common.Tools.DotNet.Pack.DotNetPackSettings()
-            {
-                OutputDirectory = Path.Combine(context.Artifacts, @"nugets"),
-                Sources = new List<string>() { Path.Combine(context.Artifacts, "nugets") },
-                NoRestore = false,
-                NoBuild = false,
-            });
-    }
 
-    private static void PackPackages(BuildContext context, string solutionToPack)
+    private static void PackNugets(BuildContext context)
     {
-        context.DotNetPack(solutionToPack, 
+        context.DotNetPack(context.PackableNugetsSlnf, 
             new Cake.Common.Tools.DotNet.Pack.DotNetPackSettings()
         {
-            OutputDirectory = Path.Combine(context.Artifacts, @"nugets"),
+            OutputDirectory = Path.Combine(context.ArtifactsNugets),
             NoRestore = true,
             NoBuild = false,
         });
@@ -206,29 +182,7 @@ public sealed class GenerateApiDocumentationTask : FrostingTask<BuildContext>
 
         if (Helpers.CanReleaseInternal())
         {
-            //GenerateApiDocumentation(context,
-            //    @$"ix.connectors\src\Ix.Connector\bin\{context.DotNetBuildSettings.Configuration}\net6.0\Ix.Connector.dll",
-            //    @"Ix.Connector");
-            //GenerateApiDocumentation(context,
-            //    @$"ix.connectors\src\Ix.Connector.S71500.WebAP\bin\{context.DotNetBuildSettings.Configuration}\net6.0\Ix.Connector.S71500.WebAPI.dll",
-            //    @"Ix.Connector.S71500.WebAPI");
-
-            //GenerateApiDocumentation(context,
-            //    @$"ix.builder\src\IX.Compiler\bin\{context.DotNetBuildSettings.Configuration}\net6.0\IX.Compiler.dll",
-            //    @"IX.Compiler");
-            //GenerateApiDocumentation(context,
-            //    @$"ix.builder\src\IX.Cs.Compiler\bin\{context.DotNetBuildSettings.Configuration}\net6.0\IX.Compiler.Cs.dll",
-            //    @"IX.Compiler.Cs");
-
-            //GenerateApiDocumentation(context,
-            //    @$"ix.abstractions\src\Ix.Abstractions\bin\{context.DotNetBuildSettings.Configuration}\net6.0\Ix.Abstractions.dll",
-            //    @"Ix.Abstractions");
-            //GenerateApiDocumentation(context,
-            //    @$"ix.blazor\src\Ix.Presentation.Blazor\bin\{context.DotNetBuildSettings.Configuration}\net6.0\Ix.Presentation.Blazor.dll",
-            //    @"Ix.Presentation.Blazor");
-            //GenerateApiDocumentation(context,
-            //    @$"ix.blazor\src\Ix.Presentation.Blazor.Controls\bin\{context.DotNetBuildSettings.Configuration}\net6.0\Ix.Presentation.Blazor.Controls.dll",
-            //    @"Ix.Presentation.Blazor.Controls");
+           
         }
     }
