+++ src/core/ctrl/apax.yml
@@ -1,5 +1,5 @@
 name: "@ix-ax/ix.framework.core"
-version   : '0.1.0-updates-from-pku-org.1'
+version          : '0.1.0-20-complete-build-script-for-ixframework.1'
 type: lib
 targets:
   - "1500"
