+++ src/core/src/ix.framework.core.blazor/ix.framework.core.blazor.csproj
@@ -21,8 +21,8 @@
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
-    <PackageReference Include="Ix.Presentation.Blazor" Version="0.13.0-alpha.32" />
-    <PackageReference Include="Ix.Presentation.Blazor.Controls" Version="0.13.0-alpha.32" />
+    <PackageReference Include="Ix.Presentation.Blazor" Version="0.13.3-alpha.53" />
+    <PackageReference Include="Ix.Presentation.Blazor.Controls" Version="0.13.3-alpha.53" />
     <PackageReference Include="Microsoft.AspNetCore.Components.Web" Version="7.0.2" />
   </ItemGroup>
