+++ src/core/src/ix.framework.core/ix.framework.core.csproj
@@ -10,7 +10,7 @@
 		  <PrivateAssets>all</PrivateAssets>
 		  <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
 		</PackageReference>
-		<PackageReference Include="Ix.Connector" Version="0.13.0-alpha.32" />
+		<PackageReference Include="Ix.Connector" Version="0.13.3-alpha.53" />
 	</ItemGroup>
 
 	<ItemGroup>
