+++ src/integrations/integration.blazor/integration.blazor.csproj
@@ -33,7 +33,7 @@
 		  <PrivateAssets>all</PrivateAssets>
 		  <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
 		</PackageReference>
-		<PackageReference Include="Ix.Presentation.Blazor.Controls" Version="0.13.0-alpha.32" />
+		<PackageReference Include="Ix.Presentation.Blazor.Controls" Version="0.13.3-alpha.53" />
 	</ItemGroup>
 	<!--REF_PCKG_END-->
 	<ItemGroup>
