+++ src/workspace.ix.framework.code-workspace
@@ -1,12 +1,12 @@
 {
 	"folders": [
 		{
-			"name": "core",
-			"path": "core\\ctrl"
-		},
-		{
 			"name": "integrations",
 			"path": "integrations\\ctrl"
+		},
+		{
+			"name": "core",
+			"path": "core\\ctrl"
 		}
 	],
 	"settings": {}
