+++ test.ps1
@@ -1,0 +1,4 @@
+# run build
+
+dotnet run --project cake/Build.csproj --do-test true --do-pack true --test-level 10
+exit $LASTEXITCODE;
